Remove bv32.rs
