+++ src/flux_support/mod.rs
@@ -1,2 +1,1 @@
-pub mod bv32;
 pub mod rmap;
